backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic-settings
cloud-sql-python-connector
google-auth
google-api-python-client
requests
pyyaml
pandas
openpyxl
pyodbc
cryptography
python-multipart
